Flask_Twitter_Sentiment_Analysis_Service/api/requirements.txt
flask==1.0.2
pytest-flask==0.10.0
pytest-mock==1.10.0